fix bug when saving images, change which experiments are used
There was a bug with function used to plot the images in the
visualization notebook. The function receveid an extra argument as the
file name for the plot to be saved, but I automated it such that the
file name is determined by the title of the plot, the column (speed,
position, acceleration), and the number of the experiment. This makes
things easier.

Additionally, I changed the experiments used for the block on slope
analysis. We were using data from experiments 1,2,3,4,5 but now are
using from the experiments 3,4,5,6,8 because they fit better in our
model.